feat(QDOG-58): Create add pet form page
Rebase this branch and update some imports to follow the format in ts config]
Move all data to constants.ts
