+++ src/app/new-pet/page.tsx
@@ -4,17 +4,17 @@ import { Box, Button, Group, Select, TextInput } from "@mantine/core";
 import { useEffect, useState } from "react";
 import { useForm } from "@mantine/form";
 import { DatePickerInput } from "@mantine/dates";
-import { sexOptions, animalGroupOptions } from "../../data/pet.ts";
-import { basicOptions } from "../../data/general.ts";
-import { defaultDate } from "../../data/date.ts";
+import { sexOptions, animalGroupOptions } from "~data/constants";
+import { basicOptions } from "~data/constants";
+import { defaultDate } from "~data/constants";
 import {
     validateImage,
     validateDateValue,
     validateSelectedAnimalGroup,
     validateSelectedSex,
     validateSelectedSpayedOrNeutered,
-} from "../../util/validation/validate-newpet.ts";
-import { validateStringValue } from "~/util/validation/validation.ts";
+} from "~util/validation/new-pet.ts";
+import { validateStringValue } from "~util/validation/validation.ts";
 import styles from "./page.module.scss";
 import placeholderImage from "../../../public/placeholder.jpg"; // https://media.istockphoto.com/id/1147544807/vector/thumbnail-image-vector-graphic.jpg?s=612x612&w=0&k=20&c=rnCKVbdxqkjlcs3xH87-9gocETqpspHFXu5dIGB4wuM=
 import { useFileDialog } from "@mantine/hooks";
