+++ src/data/constants.ts
@@ -16,3 +16,23 @@ export const provinces = [
     { value: "SK", label: "SK" },
     { value: "YT", label: "YT" },
 ];
+
+export const basicOptions: string[] = ["Yes", "No", "Unknown"];
+
+export const defaultDate: Date = new Date(0);
+
+export const todayDate: Date = new Date();
+
+export const animalGroupOptions: string[] = [
+    "Small animal (e.g. cat, rabbit, and mouse)",
+    "Farm (e.g. chicken, pig, and cow)",
+    "Equine (e.g. horse)",
+    "Bird",
+    "Reptile",
+    "Amphibian",
+    "Fish",
+    "Invertebrate",
+    "Other",
+];
+
+export const sexOptions: string[] = ["Male", "Female", "Unknown"];
