--- src/data/date.ts
@@ -1,3 +1,0 @@
-export const defaultDate: Date = new Date(0);
-
-export const todayDate: Date = new Date();
