--- src/data/general.ts
@@ -1,1 +1,0 @@
-export const basicOptions: string[] = ["Yes", "No", "Unknown"];
