--- src/data/pet.ts
@@ -1,13 +1,0 @@
-export const animalGroupOptions: string[] = [
-    "Small animal (e.g. cat, rabbit, and mouse)",
-    "Farm (e.g. chicken, pig, and cow)",
-    "Equine (e.g. horse)",
-    "Bird",
-    "Reptile",
-    "Amphibian",
-    "Fish",
-    "Invertebrate",
-    "Other",
-];
-
-export const sexOptions: string[] = ["Male", "Female", "Unknown"];
